pyproject.toml
[project]
name = "clearml-hydra-template"
version = "0.1.0"
description = "A template project for integrating ClearML experiment tracking with Hydra configuration management"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["clearml", "hydra", "mlops", "experiment-tracking", "configuration-management"]
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "clearml>=1.16.4",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/clearml-hydra-template"
Repository = "https://github.com/yourusername/clearml-hydra-template"
Issues = "https://github.com/yourusername/clearml-hydra-template/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "configs"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
